fix: Don't store custom waypoints and found chests in MainMapFeature
